Reorganizing after model, view, presenter format
