+++ src/model/store.jsx
@@ -1,0 +1,16 @@
+import { configureStore } from "@reduxjs/toolkit";
+import searchReducer from "./searchSlice";
+import sidebarReducer from "./sidebarSlice"
+import userPageReducer from "./userPageSlice"
+import movieListReducer from "./movieListSlice"
+import selectedMovieReducer from "./inspectMovieSlice"
+
+export const store = configureStore({
+    reducer: {
+        results: searchReducer,
+        sidebar: sidebarReducer,
+        movieList: movieListReducer,
+        userPage: userPageReducer,
+        selectedMovie: selectedMovieReducer,
+    }
+})
