+++ src/presenter/css/Sidebar.css
