+++ src/view/StreamingButtons.jsx
